+++ suppaftp/src/async_ftp/async_std_ftp/tls/native_tls.rs
@@ -10,7 +10,7 @@ use async_std::net::TcpStream;
 use async_trait::async_trait;
 use pin_project::pin_project;
 
-use super::{AsyncTlsConnector, AsyncTlsStream};
+use super::{AsyncStdTlsStream, AsyncTlsConnector};
 use crate::{FtpError, FtpResult};
 
 #[derive(Debug)]
@@ -85,7 +85,7 @@ impl Write for AsyncNativeTlsStream {
     }
 }
 
-impl AsyncTlsStream for AsyncNativeTlsStream {
+impl AsyncStdTlsStream for AsyncNativeTlsStream {
     type InnerStream = TlsStream<TcpStream>;
 
     fn get_ref(&self) -> &TcpStream {